move recording types to voiceRecording file
